+++ TombEngine/Objects/TR4/Entity/tr4_troops.cpp
@@ -41,11 +41,12 @@ namespace TEN::Entities::TR4
 		TROOP_STATE_FLASHED = 17
 	};
 
-	// TODO
 	enum TroopAnim
 	{
 
 	};
+
+	constexpr auto TROOPS_SHOT_DAMAGE = 23;
 
 	void InitializeTroops(short itemNumber)
 	{
@@ -130,6 +131,12 @@ namespace TEN::Entities::TR4
 					item->Animation.ActiveState = TROOP_STATE_DEATH;
 					item->Animation.FrameNumber = GetAnimData(item).frameBase;
 				}
+
+				item->ItemFlags[FINAL_SHOT_FLAG_INDEX] = Random::GenerateInt(1, FINAL_SHOT_COUNT);
+			}
+			else
+			{
+				PerformFinalAttack(*item, TroopsBite1, 8, 19, TROOPS_SHOT_DAMAGE, SFX_TR3_OIL_SMG_FIRE);
 			}
 		}
 		else
@@ -371,7 +378,7 @@ namespace TEN::Entities::TR4
 				}
 				else
 				{
-					ShotLara(item, &AI, TroopsBite1, joint0, 23);
+					ShotLara(item, &AI, TroopsBite1, joint0, TROOPS_SHOT_DAMAGE);
 					creature->MuzzleFlash[0].Bite = TroopsBite1;
 					creature->MuzzleFlash[0].Delay = 2;
 					creature->Flags = 5;
@@ -433,7 +440,7 @@ namespace TEN::Entities::TR4
 				}
 				else
 				{
-					ShotLara(item, &AI, TroopsBite1, joint0, 23);
+					ShotLara(item, &AI, TroopsBite1, joint0, TROOPS_SHOT_DAMAGE);
 					creature->MuzzleFlash[0].Bite = TroopsBite1;
 					creature->MuzzleFlash[0].Delay = 2;
 					creature->Flags = 5;